+++ apps/storybook/src/components/orderlyProvider/index.tsx
@@ -1,18 +1,18 @@
 import { FC, ReactNode } from "react";
 import { onStorybookRounteChange } from "../../hooks/useStorybookNav";
-import { LocaleProvider } from "./localeProvider";
 import { OrderlyAppRootProvider } from "./orderlyAppProvider";
+import { OrderlyLocaleProvider } from "./orderlyLocaleProvider";
 import { RouteProvider } from "./rounteProvider";
 import { WalletConnectorProvider } from "./walletConnectorProvider";
 
 export const OrderlyProvider: FC<{ children: ReactNode }> = (props) => {
   return (
     <RouteProvider onRouteChange={onStorybookRounteChange}>
-      <LocaleProvider>
+      <OrderlyLocaleProvider>
         <WalletConnectorProvider>
           <OrderlyAppRootProvider>{props.children}</OrderlyAppRootProvider>
         </WalletConnectorProvider>
-      </LocaleProvider>
+      </OrderlyLocaleProvider>
     </RouteProvider>
   );
 };