+++ packages/i18n/src/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from "./useTranslation";
+export * from "./useLocaleCode";
+export * from "./useRegisterExternalResources";