+++ packages/trading-points/src/components/ranking/generalRanking/index.ts
@@ -1,5 +1,0 @@
-export {
-  GeneralRankingWidget,
-  type GeneralRankingWidgetProps,
-} from "./generalRanking.widget";
-export { useGeneralRankingScript } from "./generalRanking.script";