+++ packages/ui-tradingview/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @orderly.network/ui-tradingview
 
+## 2.12.1
+
+### Patch Changes
+
+- publish
+- 75b87bc: publish
+- Updated dependencies
+- Updated dependencies [75b87bc]
+  - @orderly.network/react-app@2.12.1
+  - @orderly.network/hooks@2.12.1
+  - @orderly.network/i18n@2.12.1
+  - @orderly.network/net@2.12.1
+  - @orderly.network/types@2.12.1
+  - @orderly.network/ui@2.12.1
+  - @orderly.network/ui-connector@2.12.1
+  - @orderly.network/ui-positions@2.12.1
+  - @orderly.network/utils@2.12.1
+
+## 2.12.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react-app@2.12.1-alpha.0
+  - @orderly.network/hooks@2.12.1-alpha.0
+  - @orderly.network/i18n@2.12.1-alpha.0
+  - @orderly.network/net@2.12.1-alpha.0
+  - @orderly.network/types@2.12.1-alpha.0
+  - @orderly.network/ui@2.12.1-alpha.0
+  - @orderly.network/ui-connector@2.12.1-alpha.0
+  - @orderly.network/ui-positions@2.12.1-alpha.0
+  - @orderly.network/utils@2.12.1-alpha.0
+
 ## 2.12.0
 
 ### Minor Changes